% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleImputation.R
\name{MultipleImputation}
\alias{MultipleImputation}
\title{Impute missing values using mice package}
\usage{
MultipleImputation(
  dataset,
  no_pred_list,
  high_corr_list,
  low_pred_list,
  skip_list,
  creat_list,
  fat_list,
  log2_list,
  m,
  seed,
  path,
  file_name
)
}
\arguments{
\item{dataset}{A dataframe containing missing values}

\item{no_pred_list}{A character vector containing names of the variables to be skipped from imputation}

\item{high_corr_list}{A character vector containing names of the highly correlated variables to be treated specially}

\item{low_pred_list}{A character vector containing names of the variables with low prediction potential, to be treated specially}

\item{skip_list}{A character vector containing names of the variables to be skipped from imputation but not from prediction}

\item{creat_list}{A character vector containing names of the variables that were creatinine adjusted (passive imputation)}

\item{fat_list}{A character vector containing names of the variables that were total fat percentage adjusted (passive imputation)}

\item{log2_list}{List variables not adjusted for creatinine or fat but log2 transformed}

\item{m}{Number of imputations}

\item{seed}{Seed value}

\item{path}{A string with a path to save the file}

\item{file_name}{A string with a file name}
}
\value{
A mids object
}
\description{
Impute missing values using mice package
}
